+++ src/main/java/com/agustinbravop/clinica_odontologica/model/Domicilio.java
@@ -4,7 +4,7 @@ import javax.persistence.Entity;
 
 @Entity
 public class Domicilio {
-    private Integer id;
+    private Long id;
     private String calle;
     private String numero;
     private String localidad;
@@ -13,7 +13,7 @@ public class Domicilio {
     public Domicilio() {
     }
 
-    public Domicilio(Integer id, String calle, String numero, String localidad, String provincia) {
+    public Domicilio(Long id, String calle, String numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
@@ -27,11 +27,11 @@ public class Domicilio {
         this.provincia = provincia;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
