+++ src/main/java/com/agustinbravop/clinica_odontologica/model/Turno.java
@@ -5,12 +5,12 @@ import java.util.Date;
 
 @Entity
 public class Turno {
-    private Integer id;
+    private Long id;
     private Paciente paciente;
     private Odontologo odontologo;
     private Date date; /*2021-09-10*/
 
-    public Turno(Integer id, Paciente paciente, Odontologo odontologo, Date date) {
+    public Turno(Long id, Paciente paciente, Odontologo odontologo, Date date) {
         this.id = id;
         this.paciente = paciente;
         this.odontologo = odontologo;
@@ -20,11 +20,11 @@ public class Turno {
     public Turno() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
